MiraquaOfficial/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
requests-cache==1.1.1
openmeteo-requests==1.2.0
retry-requests==2.0.0
pandas==2.2.2
numpy==1.26.4
supabase==1.0.3
python-dateutil==2.9.0.post0
timezonefinder==6.3.0
google-generativeai==0.4.1
gunicorn==21.2.0
pytest==7.4.0
jsonschema==4.22.1